+++ src/app/api/doubts/deleteDoubt/[id]/route.ts
@@ -1,4 +1,5 @@
 import { deleteAnyDoubt } from "@/actions/doubts";
+import getCurrentUser from "@/actions/getCurrentUser";
 import { NextRequest, NextResponse } from "next/server";
 
 export async function DELETE(request: NextRequest,{params}:{
@@ -7,6 +8,10 @@ export async function DELETE(request: NextRequest,{params}:{
   const { id } = params;
 
   try {
+    const currentUser = await getCurrentUser();
+    if (!currentUser || currentUser.role === "STUDENT") {
+      return NextResponse.json({ error: "Unauthorized" }, { status: 400 });
+    }
     await deleteAnyDoubt(id);
     return NextResponse.json({ id });
   } catch (e: any) {
